modify error handling and rename functions to use library injection script

+++ src/zone_infos.c
@@ -6,7 +6,7 @@
 /*   By: fkoehler <fkoehler@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/09/08 17:22:05 by fkoehler          #+#    #+#             */
-/*   Updated: 2017/09/26 17:00:34 by fkoehler         ###   ########.fr       */
+/*   Updated: 2017/09/27 16:28:36 by fkoehler         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,9 +24,9 @@ char		*get_zone_type_str(t_size_type type)
 
 t_size_type	get_zone_type(size_t size)
 {
-	if (size == TINY_SIZE)
+	if (size == (size_t)TINY_SIZE)
 		return (TINY);
-	else if (size == SMALL_SIZE)
+	else if (size == (size_t)SMALL_SIZE)
 		return (SMALL);
 	else
 		return (LARGE);
